engines/fr/Previous/requirements.txt
facenet-pytorch
numpy
opencv-python-headless
Pillow
matplotlib
tqdm